refactor(did-webvh): split update.rs into submodule directory
The vta-service/src/operations/did_webvh/update.rs (~2,725 lines on
P6) is now a module directory:
- options.rs (request/result types + DerivedWebvhKey)
- errors.rs (UpdateDidWebvhError + AppError mapping)
- validate.rs (document shape checks)
- keys.rs (BIP-32 derivation, hashing, install/lookup)
- legacy.rs (pre-webvh_keys-convention fallbacks)
- state.rs (jsonl <-> DIDWebVHState round-trip + record lookup)
- orchestrator.rs (update_did_webvh - including P6's optimistic
  concurrency + pre-rotation-aware signing key selection)
- rotate.rs (rotate_did_webvh_keys)
- mod.rs (re-exports)

No public-API change; no behavior change. Every existing test passes.

Refs #6 in the workspace audit.
